Handle rising bubbles like pendent drops.
Rising bubbles could previously be handled by either flipping the image
upside-down prior to calling the plugin, or by adding 180 degres to
gravity to make it point upwards. In the latter case the initial
parameter guessing would however fail. Now the plugin attempts to detect
the capillary tube to know which way the drop points. Internally the
right/leftBorder arrays are filled with the capillary tube side always
at y=0, so that the parameter guessing algorithm remains practically
unchanged. The Contour parameter tip_y is however the real on-screen
coordinate.

Other minor changes:
- crop ROI to image if needed
- correct strange formula for number of integration steps (nPoints)
  in calculateProfile()
- the intersection detection in calcFitDistance() is now more robust,
  and relies only on the positive orientation of the contour