Minimise mean fit distance, print as RMS value
Until now version 2.0.x (x<4) minimised the sum of the squared distances
between theoretical and detected contours. To provide an easier
interpretation of the fit distance for the user, we switch to using the
mean square distance (i.e. we divide by the number of contour points on
which the distance was calculated) and print the square root of that
value both in the log window and the result table. This value is in
pixel units.